Fixed BinaryWriter.
